backend/requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coverage==6.4.2
Deprecated==1.2.13
distlib==0.3.5
dj-database-url==1.0.0
Django==4.1
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-debug-toolbar==3.5.0
django-environ==0.8.1
django-filter==22.1
django-google-maps==0.13.0
django-jinja==2.10.2
django-modeltranslation==0.18.4
django-rosetta==0.9.8
django-utils-six==2.0
djangorestframework==3.13.1
filelock==3.8.0
flower==1.1.0
gunicorn==20.1.0
humanize==4.2.3
idna==3.3
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.1
packaging==21.3
Pillow==9.2.0
pipenv==2022.8.24
platformdirs==2.5.2
polib==1.1.1
prometheus-client==0.14.1
prompt-toolkit==3.0.30
pyparsing==3.0.9
python-slugify==6.1.2
pytz==2022.1
redis==4.3.4
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
tornado==6.2
urllib3==1.26.11
vine==5.0.0
virtualenv==20.16.3
virtualenv-clone==0.5.7
wcwidth==0.2.5
whitenoise==6.2.0
wrapt==1.14.1
